Add initial ladders implementation
There's 10 ladders on each level, and they can only access spaces from 2 to 99, so there's no way to have a ladder on space 1 or space 100. There's no limit to the length of the ladders, so it's entirely possible to get a ladder from 2 to 99. I'll leave that up to probability though.